requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1; python_version >= '3.7'
aiohttp==3.9.3; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
asyncer==0.0.2; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
attrs==23.2.0; python_version >= '3.7'
bidict==0.23.1; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
chainlit==1.0.500; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
chevron==0.14.0
click==8.1.7; python_version >= '3.7'
dataclasses-json==0.5.14; python_version < '3.13' and python_version >= '3.7'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.108.0; python_version >= '3.8'
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1; python_version >= '3.8'
googleapis-common-protos==1.63.0; python_version >= '3.7'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.62.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.0; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
langchain==0.1.14; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-community==0.0.31; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-core==0.1.40; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-text-splitters==0.0.1; python_full_version >= '3.8.1' and python_version < '4.0'
langsmith==0.1.40; python_full_version >= '3.8.1' and python_version < '4.0'
lazify==0.4.0
literalai==0.0.401
marshmallow==3.21.1; python_version >= '3.8'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
nest-asyncio==1.6.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
opentelemetry-api==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-common==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-grpc==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-http==1.24.0; python_version >= '3.8'
opentelemetry-instrumentation==0.45b0; python_version >= '3.8'
opentelemetry-proto==1.24.0; python_version >= '3.8'
opentelemetry-sdk==1.24.0; python_version >= '3.8'
opentelemetry-semantic-conventions==0.45b0; python_version >= '3.8'
orjson==3.10.0; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
protobuf==4.25.3; python_version >= '3.8'
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python-engineio==4.9.0; python_version >= '3.6'
python-graphql-client==0.4.3
python-multipart==0.0.9; python_version >= '3.8'
python-socketio==5.11.2; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
setuptools==69.2.0; python_version >= '3.8'
simple-websocket==1.0.0; python_version >= '3.6'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.29; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
syncer==2.0.3
tenacity==8.2.3; python_version >= '3.7'
tomli==2.0.1; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
typing-inspect==0.9.0
uptrace==1.22.0; python_version >= '3.7'
urllib3==2.2.1; python_version >= '3.8'
uvicorn==0.25.0; python_version >= '3.8'
watchfiles==0.20.0; python_version >= '3.7'
websockets==12.0; python_version >= '3.8'
wrapt==1.16.0; python_version >= '3.6'
wsproto==1.2.0; python_full_version >= '3.7.0'
yarl==1.9.4; python_version >= '3.7'
zipp==3.18.1; python_version >= '3.8'
